refactor: 컴포즈 프리뷰 개선 및 코드 정리
