+++ app/src/main/java/woowacourse/payments/ui/cardcreate/CreateCardActivity.kt
@@ -3,10 +3,12 @@ package woowacourse.payments.ui.cardcreate
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
+import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
@@ -16,6 +18,7 @@ import woowacourse.payments.ui.cards.CardsActivity
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
 class CreateCardActivity : ComponentActivity() {
+    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
