+++ app/src/main/java/woowacourse/payments/ui/cardcreate/CreateCardScreen.kt
@@ -7,13 +7,10 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.ui.components.PaymentCard
-import woowacourse.payments.ui.model.paymentCardUiModelSample
 
 private val ScreenAppBarSpacing = 14.dp
 private val ScreenSectionSpacing = 40.dp
