+++ app/src/main/java/woowacourse/payments/ui/cards/NonCardsSection.kt
@@ -9,10 +9,13 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import woowacourse.payments.R
 import woowacourse.payments.ui.components.PaymentCreateCard
 
 @Composable
@@ -24,7 +27,7 @@ fun NonCardsSection(onAddClick: () -> Unit, modifier: Modifier = Modifier) {
     ) {
         Spacer(Modifier.height(32.dp))
         Text(
-            "새로운 카드를 등록해주세요",
+            stringResource(R.string.guide_add_new_card),
             style = MaterialTheme.typography.bodyLarge.copy(
                 fontSize = 18.sp,
                 fontWeight = FontWeight.W700
@@ -33,4 +36,10 @@ fun NonCardsSection(onAddClick: () -> Unit, modifier: Modifier = Modifier) {
         )
         PaymentCreateCard(onAddClick)
     }
+}
+
+@Preview
+@Composable
+fun NonCardsSectionPreview() {
+    NonCardsSection({})
 }
