+++ app/src/main/java/woowacourse/payments/ui/components/PaymentCardChip.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.ui.theme.YellowCB
 
@@ -22,4 +23,10 @@ fun PaymentCardChip() {
                     shape = RoundedCornerShape(4.dp),
                 ),
     )
+}
+
+@Preview
+@Composable
+fun PaymentCardChipPreview() {
+    PaymentCardChip()
 }
