+++ app/src/main/java/woowacourse/payments/ui/components/PaymentCardContent.kt
@@ -10,10 +10,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.em
 import androidx.compose.ui.unit.sp
 import woowacourse.payments.ui.model.PaymentCardUiModel
+import woowacourse.payments.ui.preview.paymentCardUiModelSample
 
 @Composable
 fun PaymentCardContent(paymentCardUiModel: PaymentCardUiModel, modifier: Modifier = Modifier) {
@@ -46,4 +48,10 @@ fun PaymentCardContent(paymentCardUiModel: PaymentCardUiModel, modifier: Modifie
             textAlign = TextAlign.Right
         )
     }
+}
+
+@Preview
+@Composable
+fun PaymentCardContentPreview() {
+    PaymentCardContent(paymentCardUiModelSample)
 }
