+++ app/src/main/java/woowacourse/payments/ui/model/PaymentCardUiModel.kt
@@ -29,13 +29,3 @@ data class PaymentCardUiModel(
         cardExpiry.chunked(2)
             .joinToString(separator)
 }
-
-val paymentCardUiModelSample =
-    PaymentCardUiModel("1234567812345678", "1225", "빰".repeat(30), "1234")
-
-val paymentCardUiModelSamples =
-    listOf(
-        paymentCardUiModelSample,
-        paymentCardUiModelSample,
-        paymentCardUiModelSample
-    )
